DeepSeek Language Model Class

Language model implementation for DeepSeek's chat completions API. Inherits from OpenAI model but adds support for DeepSeek-specific features like reasoning content extraction from deepseek-reasoner model.

Super classes

aisdk::LanguageModelV1 -> aisdk::OpenAILanguageModel -> DeepSeekLanguageModel

Methods

Public methods

DeepSeekLanguageModel$do_generate()

DeepSeekLanguageModel$clone()

Inherited methods

aisdk::LanguageModelV1$generate()

aisdk::LanguageModelV1$stream()

aisdk::OpenAILanguageModel$do_stream()

aisdk::OpenAILanguageModel$format_tool_result()

aisdk::OpenAILanguageModel$get_config()

aisdk::OpenAILanguageModel$get_history_format()

aisdk::OpenAILanguageModel$initialize()

Method do_generate()

Generate text (non-streaming).

Usage

DeepSeekLanguageModel$do_generate(params)

Arguments

params

A list of call options including messages, temperature, etc.

Returns

A GenerateResult object.

Method clone()

The objects of this class are cloneable with this method.

Usage

DeepSeekLanguageModel$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.